Domain, handlers, service clients

+++ src/Pacco.Services.Orders.Application/Commands/Handlers/DeleteOrderHandler.cs
@@ -1,7 +1,9 @@
 using System.Threading.Tasks;
 using Convey.CQRS.Commands;
 using Pacco.Services.Orders.Application.Events;
+using Pacco.Services.Orders.Application.Exceptions;
 using Pacco.Services.Orders.Application.Services;
+using Pacco.Services.Orders.Core.Exceptions;
 using Pacco.Services.Orders.Core.Repositories;
 
 namespace Pacco.Services.Orders.Application.Commands.Handlers
@@ -19,6 +21,17 @@ namespace Pacco.Services.Orders.Application.Commands.Handlers
 
         public async Task HandleAsync(DeleteOrder command)
         {
+            var order = await _orderRepository.GetAsync(command.Id);
+            if (order is null)
+            {
+                throw new OrderNotFoundException(command.Id);
+            }
+
+            if (!order.CanBeDeleted())
+            {
+                throw new CannotDeleteOrderException(command.Id);
+            }
+
             await _orderRepository.DeleteAsync(command.Id);
             await _messageBroker.PublishAsync(new OrderDeleted(command.Id));
         }
